Time: 12 ms (42.33%) | Memory: 35.5 MB (74.32%) - LeetSync

+++ 1016-subarray-sums-divisible-by-k/subarray-sums-divisible-by-k.cpp
@@ -3,14 +3,13 @@ public:
     int subarraysDivByK(vector<int>& nums, int k) {
         int n = nums.size();
 
-        unordered_map<long long,int> mpp;
+        unordered_map<int,int> mpp;
         int res = 0;
-        long long sum = 0;
+        int sum = 0;
         for(int i = 0; i<n; i++) {
             sum = (sum + nums[i]) % k;
             int val = (sum + k) % k;
             if(val == 0) res++;
-        
             if(mpp.find(val) != mpp.end()) res += mpp[val];
             mpp[val]++;
         }
